+++ src/main/java/com/nephest/battlenet/sc2/model/local/ladder/dao/LadderSearchDAO.java
@@ -219,7 +219,7 @@ public class LadderSearchDAO
                         conversionService.convert(rs.getInt("team.team_type"), TeamType.class)
                     ),
                     conversionService.convert(rs.getInt("team.tier_type"), LeagueTier.LeagueTierType.class),
-                    rs.getLong("team.division_id"),
+                    rs.getInt("team.division_id"),
                     idDec == null ? null : idDec.toBigInteger(),
                     rs.getLong("team.rating"),
                     rs.getInt("team.wins"), rs.getInt("team.losses"), rs.getInt("team.ties"),
